io error

+++ AnimationNoIKHandle.py
@@ -48,15 +48,19 @@ while(simulationStatus != 'STOP'):
                     + 'HEAD' + ' ' + str(babyHeadCoord[0]) + ' ' + str(babyHeadCoord[1]) + ' ' + str(babyHeadCoord[2]))
     print "read baby coordinates"
     ##
-    win32file.WriteFile(babyBrain, bytearray(messageToSend, 'utf-8'))
-    print "sent message to baby"
-    win32file.WriteFile(mommyBrain, bytearray(messageToSend, 'utf-8'))
-    print "sent message to mommy"
+    if(babySimulationStatus != 'STOP'):
+        win32file.WriteFile(babyBrain, bytearray(messageToSend, 'utf-8'))
+        print "sent message to baby"
+    if(mommySimulationStatus != 'STOP'):
+        win32file.WriteFile(mommyBrain, bytearray(messageToSend, 'utf-8'))
+        print "sent message to mommy"
     ##
-    babyMessage = win32file.ReadFile(babyBrain, 16000)
-    print "read message from baby ", babyMessage
-    motherMessage = win32file.ReadFile(mommyBrain, 16000)
-    print "read message from mommy ", motherMessage
+    if(babySimulationStatus != 'STOP'):
+        babyMessage = win32file.ReadFile(babyBrain, 16000)
+        print "read message from baby ", babyMessage
+    if(mommySimulationStatus != 'STOP'):
+        motherMessage = win32file.ReadFile(mommyBrain, 16000)
+        print "read message from mommy ", motherMessage
     bMessAsList = babyMessage[1].split(' ')
     mMessAsList = motherMessage[1].split(' ')
     print "mMessAsList: " + str(mMessAsList)
